benchmarks/requirements.txt
Django==1.7.7
djangorestframework==3.1.1
marshmallow==1.2.4
.
